+++ philo/time_utils.c
@@ -12,21 +12,21 @@
 
 #include "philosophers.h"
 
-void    ft_usleep(long long microseconds)
+void	ft_usleep(long microseconds)
 {
-    long long    start;
-    long long    target_time;
+	long	start;
+	long	target_time;
 
-    start = get_current_time();
-    target_time = start + (microseconds / 1000);
-    while (get_current_time() < target_time)
-        usleep(100);
+	start = get_current_time();
+	target_time = start + (microseconds / 1000);
+	while (get_current_time() < target_time)
+		usleep(100);
 }
 
-long long    get_current_time(void)
+long	get_current_time(void)
 {
-    struct timeval    tv;
+	struct timeval	tv;
 
-    gettimeofday(&tv, NULL);
-    return ((tv.tv_sec * 1000LL) + (tv.tv_usec / 1000));
+	gettimeofday(&tv, NULL);
+	return ((tv.tv_sec * 1000LL) + (tv.tv_usec / 1000));
 }
